+++ back-end/src/auth/utils/auth.security.ts
@@ -1,4 +1,4 @@
-import { generateRandomBuffer, hashPassword } from 'src/common/helpers/string.helpes';
+import { generateRandomBuffer, hashPassword } from '@/common/helpers/string.helper';
 
 export const credentialsPassword = async (password: string) => {
 	const salt = generateRandomBuffer(32);
