+++ back-end/src/common/helpers/string.helper.ts
